refactor(lib, hicn-light, vpp, hiperf): HICN-723  - move infra data structure into the shared lib  - new packet cache using double hashing and lookup on prefix suffix  - testing updates  - authenticated requests using interest manifests

+++ libtransport/includes/hicn/transport/utils/chrono_typedefs.h
@@ -79,7 +79,7 @@ class Time {
  public:
   using Clock = T;
   using TimePoint = typename Clock::time_point;
-  using Rep = double;
+  using Rep = uint64_t;
   using Seconds = std::chrono::duration<Rep>;
   using Milliseconds = std::chrono::duration<Rep, std::milli>;
   using Microseconds = std::chrono::duration<Rep, std::micro>;
